Alters flow on unix and osx
By making OpenUrlWindows not call OpenUrlRaw, it becomes possible to run a shell on windows, but target the necessary executable directly on other OSses, without too much duplication.
Of course, the quality of the code would be greatly improved if there is a way to do this directly, as we could simplify several functions more.