Moves format check to end of formurl
So that if, for whatever reason, the url base would result in an invalid url *after* including the query string, this is detected as well. has the added security benefit of being less transparent in terms of timing (i.e. time spent will be simply in terms of input size, but won't have any outliers).